running interrupts - testing what happens/7

+++ sys/main.c
@@ -43,7 +43,9 @@ void start(uint32_t* modulep, void* physbase, void* physfree)
 
 	IDTDescriptor* idt = (IDTDescriptor*)((uint64_t)IDTR.base);
 
-	printf("test for idts : [%x]\n", idt[0].offset_low|(idt[0].offset_mid<<16)|(idt[0].offset_high<<32));
+	printf("test for idts : [%x]\n", ((uint64_t)idt[0].offset_low)|
+										(((uint64_t)idt[0].offset_mid)<<16)|
+										(((uint64_t)idt[0].offset_high)<<32));
 
 
 	printf("address: [%x]\n", (uint64_t)(&interrupt_handler));
